+++ app/app.js
@@ -16,6 +16,7 @@ import { ConnectedRouter } from 'react-router-redux';
 import createHistory from 'history/createBrowserHistory';
 import 'sanitize.css/sanitize.css';
 import Bootstrap from 'bootstrap/dist/css/bootstrap.css'; // eslint-disable-line no-unused-vars
+import moment from 'moment';
 
 // Import root app
 import App from 'containers/App';
@@ -39,6 +40,8 @@ import { translationMessages } from './i18n';
 // Import CSS reset and Global Styles
 import './global-styles';
 import AppWrapper from './containers/AppWrapper/index';
+
+moment.updateLocale(moment.locale(), { invalidDate: `--` });
 
 // Create redux store with history
 const initialState = {};