+++ app/components/organisms/ListView/index.js
@@ -12,7 +12,6 @@ import queryString from 'query-string';
 import config from 'config';
 import Pagination from 'components/Pagination/Loadable';
 import ErrorMsg from 'components/molecules/ErrorMsg';
-import { ListViewWrapper } from './styled';
 import listViewMessages from './messages';
 import TableDataSet from '../TableDataSet/index';
 
@@ -76,6 +75,7 @@ class ListView extends React.PureComponent {
     const ceilConfiguration = this.props.ceilConfiguration || {
       id: value => <Link to={`${this.props.link}/${value}`}>{value}</Link>,
     };
+    const headerConfiguration = this.props.headerConfiguration || {};
 
     const page = parseInt(parsedQuery.page || 1, 10);
     const sort = parsedQuery.sort || this.props.defaultSort;
@@ -105,7 +105,7 @@ class ListView extends React.PureComponent {
       );
     }
     return (
-      <ListViewWrapper>
+      <section>
         <TableDataSet
           name={name}
           columns={columns}
@@ -114,6 +114,7 @@ class ListView extends React.PureComponent {
           sortBy={sort}
           orderBy={order}
           ceilConfiguration={ceilConfiguration}
+          headerConfiguration={headerConfiguration}
           data={
             list.data.length > config.limit ? list.data.slice(0, -1) : list.data
           }
@@ -129,7 +130,7 @@ class ListView extends React.PureComponent {
           order={order}
           nextPage={list.data.length > config.limit}
         />
-      </ListViewWrapper>
+      </section>
     );
   }
 }
@@ -142,6 +143,7 @@ ListView.propTypes = {
   columns: PropTypes.object.isRequired,
   sortingColumns: PropTypes.array.isRequired,
   ceilConfiguration: PropTypes.object,
+  headerConfiguration: PropTypes.object,
   defaultSort: PropTypes.string,
   defaultOrder: PropTypes.string,
   link: PropTypes.string.isRequired,
@@ -154,6 +156,7 @@ ListView.defaultProps = {
   defaultSort: 'id',
   defaultOrder: 'desc',
   ceilConfiguration: null,
+  headerConfiguration: null,
 };
 
 ListView.contextTypes = {