+++ app/containers/BlockexplorerDashboardPage/tests/actions.test.js
@@ -1,38 +1,66 @@
 import {
-  loadLatestNode,
+  loadTopNodes,
+  loadTopAccounts,
   loadLatestBlocks,
+  loadLatestsMessages,
   loadLatestsTransactions,
-  latestNodesLoaded,
+  topNodesLoaded,
+  topAccountsLoaded,
   latestBlocksLoaded,
+  latestMessagesLoaded,
   latestTransactionsLoaded,
 } from '../actions';
 import {
-  LOAD_LATEST_NODES,
+  LOAD_TOP_NODES,
+  LOAD_TOP_ACCOUNTS,
   LOAD_LATEST_BLOCKS,
+  LOAD_LATEST_MESSAGES,
   LOAD_LATEST_TRANSACTIONS,
-  LOAD_LATEST_NODES_SUCCESS,
+  LOAD_TOP_NODES_SUCCESS,
+  LOAD_TOP_ACCOUNTS_SUCCESS,
   LOAD_LATEST_BLOCKS_SUCCESS,
+  LOAD_LATEST_MESSAGES_SUCCESS,
   LOAD_LATEST_TRANSACTIONS_SUCCESS,
 } from '../constants';
 
 describe('BlockexplorerDashboardPage actions', () => {
   describe('Load nodes Action', () => {
-    it('has a type of LOAD_LATEST_NODES', () => {
+    it('has a type of LOAD_TOP_NODES', () => {
       const expected = {
-        type: LOAD_LATEST_NODES,
+        type: LOAD_TOP_NODES,
       };
-      expect(loadLatestNode()).toEqual(expected);
+      expect(loadTopNodes()).toEqual(expected);
     });
   });
-  describe('Latest nodes loaded Action', () => {
-    it('has a type of LOAD_LATEST_NODES_SUCCESS', () => {
+  describe('Top nodes loaded Action', () => {
+    it('has a type of LOAD_TOP_NODES_SUCCESS', () => {
       const nodes = [];
 
       const expected = {
-        type: LOAD_LATEST_NODES_SUCCESS,
+        type: LOAD_TOP_NODES_SUCCESS,
         nodes,
       };
-      expect(latestNodesLoaded(nodes)).toEqual(expected);
+      expect(topNodesLoaded(nodes)).toEqual(expected);
+    });
+  });
+
+  describe('Load accounts Action', () => {
+    it('has a type of LOAD_TOP_ACCOUNTS', () => {
+      const expected = {
+        type: LOAD_TOP_ACCOUNTS,
+      };
+      expect(loadTopAccounts()).toEqual(expected);
+    });
+  });
+  describe('Top accounts loaded Action', () => {
+    it('has a type of LOAD_TOP_ACCOUNTS_SUCCESS', () => {
+      const accounts = [];
+
+      const expected = {
+        type: LOAD_TOP_ACCOUNTS_SUCCESS,
+        accounts,
+      };
+      expect(topAccountsLoaded(accounts)).toEqual(expected);
     });
   });
 
@@ -53,6 +81,26 @@ describe('BlockexplorerDashboardPage actions', () => {
         blocks,
       };
       expect(latestBlocksLoaded(blocks)).toEqual(expected);
+    });
+  });
+
+  describe('Load messages Action', () => {
+    it('has a type of LOAD_LATEST_MESSAGES', () => {
+      const expected = {
+        type: LOAD_LATEST_MESSAGES,
+      };
+      expect(loadLatestsMessages()).toEqual(expected);
+    });
+  });
+  describe('Latest messages loaded Action', () => {
+    it('has a type of LOAD_LATEST_MESSAGES_SUCCESS', () => {
+      const messages = [];
+
+      const expected = {
+        type: LOAD_LATEST_MESSAGES_SUCCESS,
+        messages,
+      };
+      expect(latestMessagesLoaded(messages)).toEqual(expected);
     });
   });