+++ app/containers/BlockexplorerDashboardPage/tests/saga.test.js
@@ -4,18 +4,30 @@
 
 /* eslint-disable redux-saga/yield-effects */
 import { put, takeLatest } from 'redux-saga/effects';
-import defaultSaga, { getNodes, getBlocks, getTransactions } from '../saga';
+import defaultSaga, {
+  getNodes,
+  getAccounts,
+  getBlocks,
+  getMessages,
+  getTransactions,
+} from '../saga';
 import {
-  latestNodesLoaded,
-  latestNodesLoadingError,
+  topNodesLoaded,
+  topNodesLoadingError,
+  topAccountsLoaded,
+  topAccountsLoadingError,
   latestBlocksLoaded,
   latestBlocksLoadingError,
+  latestMessagesLoaded,
+  latestMessagesLoadingError,
   latestTransactionsLoaded,
   latestTransactionsLoadingError,
 } from '../actions';
 import {
+  LOAD_TOP_NODES,
+  LOAD_TOP_ACCOUNTS,
   LOAD_LATEST_BLOCKS,
-  LOAD_LATEST_NODES,
+  LOAD_LATEST_MESSAGES,
   LOAD_LATEST_TRANSACTIONS,
 } from '../constants';
 
@@ -29,7 +41,7 @@ describe('getNodes Saga', () => {
     expect(selectDescriptor).toMatchSnapshot();
   });
 
-  it('should dispatch the latestNodesLoaded action if it requests the data successfully', () => {
+  it('should dispatch the topNodesLoaded action if it requests the data successfully', () => {
     const nodes = [
       {
         id: '0001',
@@ -39,13 +51,43 @@ describe('getNodes Saga', () => {
       },
     ];
     const putDescriptor = getNodesGenarator.next(nodes).value;
-    expect(putDescriptor).toEqual(put(latestNodesLoaded(nodes)));
+    expect(putDescriptor).toEqual(put(topNodesLoaded(nodes)));
   });
 
-  it('should call the latestNodesLoadingError action if the response errors', () => {
+  it('should call the topNodesLoadingError action if the response errors', () => {
     const response = new Error('Some error');
     const putDescriptor = getNodesGenarator.throw(response).value;
-    expect(putDescriptor).toEqual(put(latestNodesLoadingError(response)));
+    expect(putDescriptor).toEqual(put(topNodesLoadingError(response)));
+  });
+});
+
+describe('getAccounts Saga', () => {
+  let getAccountsGenarator;
+
+  beforeEach(() => {
+    getAccountsGenarator = getAccounts();
+
+    const selectDescriptor = getAccountsGenarator.next().value;
+    expect(selectDescriptor).toMatchSnapshot();
+  });
+
+  it('should dispatch the topAccountsLoaded action if it requests the data successfully', () => {
+    const accounts = [
+      {
+        id: '0001-00000000-9B6F',
+      },
+      {
+        id: '0001-00000001-8B4E',
+      },
+    ];
+    const putDescriptor = getAccountsGenarator.next(accounts).value;
+    expect(putDescriptor).toEqual(put(topAccountsLoaded(accounts)));
+  });
+
+  it('should call the topAccountsLoadingError action if the response errors', () => {
+    const response = new Error('Some error');
+    const putDescriptor = getAccountsGenarator.throw(response).value;
+    expect(putDescriptor).toEqual(put(topAccountsLoadingError(response)));
   });
 });
 
@@ -76,6 +118,36 @@ describe('getBlocks Saga', () => {
     const response = new Error('Some error');
     const putDescriptor = getBlocksGenerator.throw(response).value;
     expect(putDescriptor).toEqual(put(latestBlocksLoadingError(response)));
+  });
+});
+
+describe('getMessages Saga', () => {
+  let getMessagesGenerator;
+
+  beforeEach(() => {
+    getMessagesGenerator = getMessages();
+
+    const selectDescriptor = getMessagesGenerator.next().value;
+    expect(selectDescriptor).toMatchSnapshot();
+  });
+
+  it('should dispatch the latestMessagesLoaded action if it requests the data successfully', () => {
+    const messages = [
+      {
+        id: '0001:00000001',
+      },
+      {
+        id: '0001:00000002',
+      },
+    ];
+    const putDescriptor = getMessagesGenerator.next(messages).value;
+    expect(putDescriptor).toEqual(put(latestMessagesLoaded(messages)));
+  });
+
+  it('should call the latestMessagesLoadingError action if the response errors', () => {
+    const response = new Error('Some error');
+    const putDescriptor = getMessagesGenerator.throw(response).value;
+    expect(putDescriptor).toEqual(put(latestMessagesLoadingError(response)));
   });
 });
 
@@ -114,21 +186,33 @@ describe('getTransactions Saga', () => {
 describe('defaultSaga Saga', () => {
   const saga = defaultSaga();
 
-  it('should start task to watch for LOAD_LATEST_NODES action', () => {
+  it('should start task to watch for LOAD_TOP_NODES action', () => {
+    const takeLatestDescriptor = saga.next().value;
+    expect(takeLatestDescriptor).toEqual(takeLatest(LOAD_TOP_NODES, getNodes));
+  });
+
+  it('second task should be LOAD_TOP_ACCOUNTS action', () => {
     const takeLatestDescriptor = saga.next().value;
     expect(takeLatestDescriptor).toEqual(
-      takeLatest(LOAD_LATEST_NODES, getNodes),
+      takeLatest(LOAD_TOP_ACCOUNTS, getAccounts),
     );
   });
 
-  it('second task should be LOAD_LATEST_BLOCKS action', () => {
+  it('third task should be LOAD_LATEST_BLOCKS action', () => {
     const takeLatestDescriptor = saga.next().value;
     expect(takeLatestDescriptor).toEqual(
       takeLatest(LOAD_LATEST_BLOCKS, getBlocks),
     );
   });
 
-  it('third task should be LOAD_LATEST_TRANSACTIONS action', () => {
+  it('fourth task should be LOAD_LATEST_MESSAGES action', () => {
+    const takeLatestDescriptor = saga.next().value;
+    expect(takeLatestDescriptor).toEqual(
+      takeLatest(LOAD_LATEST_MESSAGES, getMessages),
+    );
+  });
+
+  it('fifth task should be LOAD_LATEST_TRANSACTIONS action', () => {
     const takeLatestDescriptor = saga.next().value;
     expect(takeLatestDescriptor).toEqual(
       takeLatest(LOAD_LATEST_TRANSACTIONS, getTransactions),